Handle gunicorn timeouts better.
When a request times out, the gunicorn master will set SIGABRT to the
worker process. This interrupts whatever the worker is currently
running, and runs its
[signal hander](https://github.com/benoitc/gunicorn/blob/master/gunicorn/workers/base.py#L201).
This handler:

1) mark the worker as no longer alive
2) call the `worker_abort` hook, which by default does nothing
2) sys.exit(1)

This means the in-flight request just dies. It doesn't return
a response. If we had nginx, the user may get an nginx error page, but
in airlock we don't, so the user just gets connection reset browser
error.

It also doesn't log the request, and it doesn't emit any telemetry. It
means these timeouts are kind of invisible to us. All we get is the same
gunicorn worker traceback in the logs, and no indication that what url
was being handled, or what part of the request may have caused the time
out. In airlock, we don't even get a sentry error.

However, we can do much better than this. If we define our own
`worker_abort` function, it can raise the newly added RequestTimeout
exception. That will cause an exception to be raised in the code where
the request is currently stuck. This brings many benefits:

1. The traceback will show the url that was being handled, and exactly
   what the code was doing when it timed out, which really helps
   identify the problems much faster.
2. The django machinery will respond to the user with a rendered 504
   response, rather than nginx 504 or connection reset
3. Gunicorn will emit an access log for the request, which is useful
   with airlock, as we have no nginx.
4. Our otel configuration will emit telemetry for the request, which we
   can then use to measure and alert on timeouts.

This is a huge improvement on the visibility and debugability of these
time outs.

Because we raise an exception, the handle_abort function never calls
sys.exit(1). Instead, when the worker finishes handling exception, it
exits cleanly, because it set `self.alive = False`. This means there's
a small behaviour change - the worker now exits cleanly with exit code
0, rather than rudely with 1. I think this is another improvement, as it
means any actual non-zero exit codes are not hidden by the noise of
timeout's exiting with a 1.