More gracefully handle connection timeouts with gunicorn
We discovered that chrome is eagerly connecting to gunicorn, but not
sending a request through, as an optimisation. Normally, there would be
nginx or similar in front of gunicorn, but not in airlock's case.

When this happens, the gunicorn worker timesout, and puts an ugly
stacktrace in the logs, with nothing we can do about it. It happens
a fair bit, so it's quite noisy in the logs.

This change implements a bit of a hack to allow us to detect this case,
and instead of exiting with an exception traceback, we can log one line
and then exit cleanly.