Do not load content renderer templates on import
Also includes a simplyfying refactor of the template caching.

Currently, the templates for content rendering are loaded at import
time. This is before settings.py has properly been set up, which means
that the template tags available to these import-time loaded templates
does not include slippers tags. Until now that hasn't mattered, but we
want to be able to use them in content rendering templates.

The solution is to always dynamically fetch the template on use. This is
fast in prod, as django caches template loads.  It's slower when
DEBUG=True, but we're already working with that for our other templates.

This insight into how django caches template prompted a simplifying
change in the design of template caching.  Previously, we had been
manually caching it, which required some complex logic to manage in dev,
where we want to invalidate the caching if the file is changed
underneath us. Instead, now we are using django's default template
loading (cache in prod, always reload in dev), we can use it to simplify
things greatly.

We now calculate the hash of the template contents once and store it on
the template object as a custom `airlock_` prefixed attribute.

In prod, the contents of the template will never change anyway, and
django will always return the same cached Template object.

In dev, the template is always reloaded, and new Template instance
created. So we will need to recalculate the hash every time. Whilst
a little inefficient, it ensures the hash will always be correct,
picking up changes to the content automatically.

Also switched to md5, as is much faster than SHA1, and suitable for our
purposes.