Truncate long log files to last 10k by default
Configurable in settings.

For better UX, we take care to strip any partial lines that have been
mangled by the truncation. We also display a message that the file has
been truncted due to length. We use a slippers component to do this,
which is only possible because fo the previous rework of
RendererTemplate to lazy load the template.

Note: this works, but the styles for the component are not loaded, as
currently the content rendering templates are very bare bones. Will fix
this in a subsequent commit