Refactor the file content rendering to have a base template.
This ensures we have all the styles/scripts we need, and proper html
page, so we can use our components if needed, but is free from the
general styling (e.g. header) that the main base.html provides, so
suitable for including in an iframe.

The csv template already included this, but this change factors that out
into a shared base template used by both the free text and csv
templates. Along the way there's few small QoL changes

1) use the filename as the page title

2) don't add additional newlines to preformatted text. This was unoticed
   before, but now that we have a proper html template, with
   indentation, it was obviously incorrectly adding the indentation
   inside the <pre> tag.