Add a matching loading-indicator element on the index page
Makes the wait_for_htmx() in the functional tests actually work.

I think what was happening here is that on a request/group page,
there's no matching element for htmx to add the htmx-request class
to, and so the wait_for_htmx() wasn't waiting properly. We add a
span with no other classes, so it doesn't affect anything on the
page.