Wait for htmx in functional tests for tree expand/collapse
Clicking the file link in the tree doesn't always work; from tracing,
the tree element is visible, but clicking it doesn't load the expected
content into the main page. However, when that happens, it ALSO hasn't
loaded it from an earlier step (group or folder view). Make it wait for
htmx to be done on each click.