Reduce hypothesis examples to speed up functional tests
We have hypothesis tests in the functional tests of the csv viewer,
and they are quite slow. We've now run these tests many times without
errors, so it should be safe to reduce the examples for normal test
runs and just run them with the larger number of examples in the
scheduled functional test run that happens overnight.