Better error message if no files on request
We're planning to use this management command for automating the
creation of release requests for regularly run jobs. The jobs
themselves will be run manually from jobserver, and output checkers
will need to review and release the files. We're likely to get
unsubmittable release request attempts if we haven't re-run the
job since the previous request. This will make release request
with no files, so instead of the default error message, we report
the number of already released files which should make errors
easier to interpret.