Add a weekly job to create regular release requests
Allows a backend to run regular release requests (note these are
under the "weekly" jobs, but this really just a way of grouping
jobs. We can run the group on any frequency we want, and it doesn't
run by default.)

The jobs reads config from a json file located in settings.WORKDIR,
so we don't have to configure the actual workspaces in code. Any
exceptions are logged and traced, so we can create honeycomb alerts
on errors (which wil include e.g. a failure to create a release
request because all files have already been released and we haven't
rerun the job, or a release request already exists because output
checkers haven't reviewed/released the previous one yet)