Move release request creation to actions module
Reduce the output and log instead of writing to stdout.

Also raises the FileNotFound exception if we're also submitting the
request; we want to make sure that we're not creating submitted
requests with bad files. Pending ones are OK, since the author will
need to review and manually submit anyway.

Means the management command and the runjobs job can use the same
function without one calling the other.