Extend create_release_request action to allow supporting files
This means we can create an automated release request that includes
supporting files.

Supporting files must be specified by a full path, and must be
contained within one of the specified output file dirs. As groups
are generated based on the specified dirs, this ensures that
supporting files get put in the same group. We assume that there
are relatively few supporting files, so specifying them
individually should not be too onerous.